refactor: reorganize API to domain-driven structure
Reorganize the API codebase following domain-driven design principles
where each feature/domain is self-contained with its own models, schemas,
services, and endpoints.

**New Structure:**
- `app/api/v1/domains/countries/` - Country management
- `app/api/v1/domains/users/` - User management
- `app/api/v1/domains/auth/` - Authentication (login, password recovery)
- `app/api/v1/domains/home/` - Home/welcome endpoints
- `app/api/v1/domains/images/` - Image serving
- `app/api/v1/domains/entities/` - Entities API (already existed)
- `app/api/v1/shared/` - Shared code (deps, enums)

**Changes:**
- Move models, schemas, services, endpoints to domain folders
- Move `app/api/deps.py` to `app/api/v1/shared/deps/dependencies.py`
- Move enums to `app/api/v1/shared/enums/`
- Update all imports to use new domain paths
- Remove old flat structure (`models/`, `schemas/`, `services/`, `endpoints/`, `enums/`)

**Benefits:**
- Better separation of concerns
- Easier to add new domains
- Reduced merge conflicts in team environments
- Clearer file ownership and organization
