--- app/api/deps.py
@@ -1,64 +1,0 @@
-from typing import Generator
-
-from fastapi import Depends, HTTPException, status
-from fastapi.security import OAuth2PasswordBearer
-from jose import jwt
-from pydantic import ValidationError
-from sqlalchemy.orm import Session
-
-from .v1.services import user as user_crud
-from .v1.models import User as UserModel
-from .v1.schemas import TokenPayload as TokenPayloadSchema
-
-from app.core import security
-from app.core.config import settings
-from app.db.session import SessionLocal
-
-reusable_oauth2 = OAuth2PasswordBearer(
-    tokenUrl=f"/api/v{settings.API_VERSION}/auth/login"
-)
-
-
-def get_db() -> Generator:
-    try:
-        db = SessionLocal()
-        yield db
-    finally:
-        db.close()
-
-
-def get_current_user(
-    db: Session = Depends(get_db), token: str = Depends(reusable_oauth2)
-) -> UserModel:
-    try:
-        payload = jwt.decode(
-            token, settings.SECRET_KEY, algorithms=[security.ALGORITHM]
-        )
-        token_data = TokenPayloadSchema(**payload)
-    except (jwt.JWTError, ValidationError):
-        raise HTTPException(
-            status_code=status.HTTP_403_FORBIDDEN,
-            detail="Could not validate credentials",
-        )
-    user = user_crud.get(db, id=token_data.sub)
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-    return user
-
-
-def get_current_active_user(
-    current_user: UserModel = Depends(get_current_user),
-) -> UserModel:
-    if not user_crud.is_active(current_user):
-        raise HTTPException(status_code=400, detail="Inactive user")
-    return current_user
-
-
-def get_current_active_superuser(
-    current_user: UserModel = Depends(get_current_user),
-) -> UserModel:
-    if not user_crud.is_superuser(current_user):
-        raise HTTPException(
-            status_code=400, detail="The user doesn't have enough privileges"
-        )
-    return current_user
