+++ app/api/v1/domains/auth/__init__.py
@@ -1,0 +1,1 @@
+"""Auth domain - handles authentication and authorization"""
