+++ app/api/v1/domains/auth/endpoints/__init__.py
@@ -1,0 +1,3 @@
+from app.api.v1.domains.auth.endpoints.auth import router
+
+__all__ = ["router"]
