+++ app/api/v1/domains/countries/__init__.py
@@ -1,0 +1,1 @@
+"""Countries domain - manages country data"""
