+++ app/api/v1/domains/countries/endpoints/__init__.py
@@ -1,0 +1,3 @@
+from app.api.v1.domains.countries.endpoints.countries import router
+
+__all__ = ["router"]
