+++ app/api/v1/domains/countries/models/__init__.py
@@ -1,0 +1,3 @@
+from app.api.v1.domains.countries.models.country import Country
+
+__all__ = ["Country"]
