+++ app/api/v1/domains/countries/schemas/__init__.py
@@ -1,0 +1,3 @@
+from app.api.v1.domains.countries.schemas.country import Country, CountryCreate, CountryUpdate, CountryInDBBase
+
+__all__ = ["Country", "CountryCreate", "CountryUpdate", "CountryInDBBase"]
