+++ app/api/v1/domains/countries/services/__init__.py
@@ -1,0 +1,3 @@
+from app.api.v1.domains.countries.services.country import country
+
+__all__ = ["country"]
