+++ app/api/v1/domains/countries/services/country.py
@@ -1,0 +1,44 @@
+from typing import Any, Dict, List, Optional, Union
+
+from sqlalchemy.orm import Session
+
+from app.api.common.services import CRUDBaseService
+from app.api.v1.domains.countries.models.country import Country
+from app.api.v1.domains.countries.schemas.country import CountryCreate, CountryUpdate
+
+
+class CountryService(CRUDBaseService[Country, CountryCreate, CountryUpdate]):
+
+    def create(self, db: Session, *, obj_in: CountryCreate) -> Country:
+        db_obj = Country(
+            name=obj_in.name,
+            status=obj_in.status
+        )
+        db.add(db_obj)
+        db.commit()
+        db.refresh(db_obj)
+        return db_obj
+
+    def update(
+        self, db: Session, *, db_obj: Country, obj_in: Union[CountryUpdate, Dict[str, Any]]
+    ) -> Country:
+        if isinstance(obj_in, dict):
+            update_data = obj_in
+        else:
+            update_data = obj_in.model_dump(exclude_unset=True)
+        return super().update(db, db_obj=db_obj, obj_in=update_data)
+
+    def get_multi(
+        self, db: Session, *, relations: Optional[List[str]] = None
+    ) -> List[Country]:
+        """
+        Get multiple countries.
+        """
+        query = db.query(self.model)
+        if relations:
+            from sqlalchemy.orm import selectinload
+            query = query.options(*[selectinload(getattr(self.model, r)) for r in relations])
+        return query.all()
+
+
+country = CountryService(Country)
