+++ app/api/v1/domains/home/__init__.py
@@ -1,0 +1,1 @@
+"""Home domain - API welcome and info endpoints"""
