+++ app/api/v1/domains/home/endpoints/__init__.py
@@ -1,0 +1,3 @@
+from app.api.v1.domains.home.endpoints.home import router
+
+__all__ = ["router"]
