+++ app/api/v1/domains/home/endpoints/home.py
