+++ app/api/v1/domains/images/__init__.py
@@ -1,0 +1,1 @@
+"""Images domain - serves static image files"""
