+++ app/api/v1/domains/images/endpoints/__init__.py
@@ -1,0 +1,3 @@
+from app.api.v1.domains.images.endpoints.images import router
+
+__all__ = ["router"]
