+++ app/api/v1/domains/users/__init__.py
@@ -1,0 +1,1 @@
+"""Users domain - manages user authentication and authorization"""
