+++ app/api/v1/domains/users/endpoints/__init__.py
@@ -1,0 +1,3 @@
+from app.api.v1.domains.users.endpoints.users import router
+
+__all__ = ["router"]
