+++ app/api/v1/domains/users/models/__init__.py
@@ -1,0 +1,3 @@
+from app.api.v1.domains.users.models.user import User
+
+__all__ = ["User"]
