+++ app/api/v1/domains/users/schemas/__init__.py
@@ -1,0 +1,4 @@
+from app.api.v1.domains.users.schemas.user import User, UserCreate, UserUpdate, UserInDB, UserInDBBase
+from app.api.v1.domains.users.schemas.token import Token, TokenPayload
+
+__all__ = ["User", "UserCreate", "UserUpdate", "UserInDB", "UserInDBBase", "Token", "TokenPayload"]
