+++ app/api/v1/domains/users/services/__init__.py
@@ -1,0 +1,3 @@
+from app.api.v1.domains.users.services.user import user
+
+__all__ = ["user"]
