--- app/api/v1/endpoints/login.py
@@ -1,101 +1,0 @@
-from datetime import timedelta, datetime
-from typing import Any
-
-from fastapi import APIRouter, Body, Depends, HTTPException
-from fastapi.security import OAuth2PasswordRequestForm
-from sqlalchemy.orm import Session
-
-from ..services import user as user_crud
-from ..models import User as UserModel
-from ..schemas import User as UserSchema, Msg as MsgSchema, Token
-from app.api import deps
-from app.core import security
-from app.core.config import settings
-from app.core.security import get_password_hash
-from app.utils import (
-    generate_password_reset_token,
-    send_reset_password_email,
-    verify_password_reset_token,
-)
-
-router = APIRouter()
-
-
-@router.post("/login", response_model=Token)
-def login_access_token(
-        db: Session = Depends(deps.get_db), form_data: OAuth2PasswordRequestForm = Depends()
-) -> Any:
-    """
-    OAuth2 compatible token login, get an access token for future requests
-    """
-    user = user_crud.authenticate(
-        db, email=form_data.username, password=form_data.password
-    )
-
-    if user is None:
-        raise HTTPException(status_code=400, detail="Incorrect email or password")
-    elif not user_crud.is_active(user):
-        raise HTTPException(status_code=400, detail="Inactive user")
-    access_token_expires = timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
-    return {
-        "access_token": security.create_access_token(
-            user.id, expires_delta=access_token_expires
-        ),
-        "expires_at": datetime.utcnow() + access_token_expires,
-        "token_type": "Bearer",
-    }
-
-
-@router.post("/me", response_model=UserSchema)
-def test_token(current_user: UserModel = Depends(deps.get_current_user)) -> Any:
-    """
-    Test access token
-    """
-    return current_user
-
-
-# TODO: Check this
-@router.post("/password-recovery/{email}", response_model=MsgSchema)
-def recover_password(email: str, db: Session = Depends(deps.get_db)) -> Any:
-    """
-    Password Recovery
-    """
-    user = user_crud.get_by_email(db, email=email)
-
-    if not user:
-        raise HTTPException(
-            status_code=404,
-            detail="The user with this username does not exist in the system.",
-        )
-    password_reset_token = generate_password_reset_token(email=email)
-    send_reset_password_email(
-        email_to=user.email, email=email, token=password_reset_token
-    )
-    return {"msg": "Password recovery email sent"}
-
-
-@router.post("/reset-password/", response_model=MsgSchema)
-def reset_password(
-        token: str = Body(...),
-        new_password: str = Body(...),
-        db: Session = Depends(deps.get_db),
-) -> Any:
-    """
-    Reset password
-    """
-    email = verify_password_reset_token(token)
-    if not email:
-        raise HTTPException(status_code=400, detail="Invalid token")
-    user = user_crud.get_by_email(db, email=email)
-    if not user:
-        raise HTTPException(
-            status_code=404,
-            detail="The user with this username does not exist in the system.",
-        )
-    elif not user_crud.is_active(user):
-        raise HTTPException(status_code=400, detail="Inactive user")
-    hashed_password = get_password_hash(new_password)
-    user.hashed_password = hashed_password
-    db.add(user)
-    db.commit()
-    return {"msg": "Password updated successfully"}
