--- app/api/v1/enums/__init__.py
@@ -1,2 +1,0 @@
-from .types import CharacterType
-from .gender import Gender
