--- app/api/v1/enums/gender.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class Gender(str, Enum):
-    male = "Male"
-    female = "Female"
-    unknown = "Unknown"
