--- app/api/v1/enums/types.py
@@ -1,10 +1,0 @@
-from enum import Enum
-
-
-class CharacterType(str, Enum):
-    human = "Human"
-    animal = "Animal"
-    ghost = "Ghost"
-    demon = "Demon"
-    apparition = "Apparition"
-    spirit = "Spirit"
