--- app/api/v1/models/user.py
@@ -1,16 +1,0 @@
-from typing import Optional
-
-from sqlalchemy import Boolean, String
-from sqlalchemy.orm import Mapped, mapped_column
-
-from app.db.base_class import Base
-
-
-class User(Base):
-    id: Mapped[int] = mapped_column(primary_key=True, index=True)
-    full_name: Mapped[Optional[str]] = mapped_column(String, index=True, nullable=True)
-    email: Mapped[str] = mapped_column(String, unique=True, index=True, nullable=False)
-    hashed_password: Mapped[str] = mapped_column(String, nullable=False)
-    is_active: Mapped[bool] = mapped_column(Boolean(), default=True)
-    is_superuser: Mapped[bool] = mapped_column(Boolean(), default=False)
-
