--- app/api/v1/schemas/__init__.py
@@ -1,4 +1,0 @@
-from .user import User, UserCreate, UserInDB, UserUpdate, UserBase
-from .msg import Msg
-from .token import Token, TokenPayload
-from .country import Country, CountryCreate, CountryUpdate, CountryInDBBase
