--- app/api/v1/schemas/msg.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-
-
-class Msg(BaseModel):
-    msg: str
