--- app/api/v1/services/__init__.py
@@ -1,3 +1,0 @@
-from .user_service import user
-from .country_service import country
-from .image_service import image_service
