+++ app/api/v1/shared/__init__.py
@@ -1,0 +1,1 @@
+"""Shared code for v1 API - deps, enums, and common utilities"""
