+++ app/db/base.py
@@ -1,8 +1,8 @@
 # Import all the models, so that Base has them before being
 # imported by Alembic
 from app.db.base_class import Base  # noqa
-from app.api.v1.models import User  # noqa
-from app.api.v1.models import Country  # noqa
+from app.api.v1.domains.users.models.user import User  # noqa
+from app.api.v1.domains.countries.models.country import Country  # noqa
 from app.api.v1.entities.models import Category  # noqa
 from app.api.v1.entities.models import EntityType  # noqa
 from app.api.v1.entities.models import Entity  # noqa
